requirements.txt
asgiref==3.11.1
Django==6.0.3
pillow==12.2.0
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.17.1
PyYAML==6.0.2
inflection==0.5.1
